code cleanup

+++ firmware/Greeting.cpp
@@ -10,8 +10,8 @@ void bounce(LedController& leds){
     6,4, 7,3, 8,2, 9,1, 8,0, 7,1, 6,2, 5,3,
     4,4
   };
-  for(int i=0; i<33; ++i){
-//   for(int i=0; i<6; ++i){
+  for(int8_t i=0; i<33; ++i){
+//   for(int8_t i=0; i<6; ++i){
     leds.fade(1);
     leds.setCross(coords[i*2], 7 - coords[i*2+1], 0xff);
     delay(80);
@@ -19,7 +19,7 @@ void bounce(LedController& leds){
 }
 
 void doubleCross(LedController& leds){
-  for(int i=0; i<8; ++i){
+  for(int8_t i=0; i<8; ++i){
     leds.fade(1);
     leds.setCross(i+2, i, 0xff);
     leds.setCross(i, 7-i, 0xff);
@@ -29,7 +29,7 @@ void doubleCross(LedController& leds){
 
 void rain(LedController& leds){
   // greeting
-  for(int i=0; i<48; ++i){
+  for(int8_t i=0; i<48; ++i){
     leds.fade(1);
     leds.setCross(i%10, 7-(i%8), 0xff);
     delay(60);
@@ -37,14 +37,14 @@ void rain(LedController& leds){
 }
 
 void fade(LedController& leds){
-  for(int i=0; i<8; ++i){
+  for(int8_t i=0; i<8; ++i){
     delay(160);
     leds.fade(1);
   }    
 }
 
 void explode(LedController& leds){
-  for(int i=0; i<8; ++i){
+  for(int8_t i=0; i<8; ++i){
     delay(60);
     leds.brighten(1);
   }
